+++ examples/basic/index.ts
@@ -1,1 +1,5 @@
-// import {  } from 'tseudo';
+import tseudo from 'tseudo';
+
+const user = tseudo.Generators.identification.person.fullName();
+
+console.log(user);