+++ src/core/register.ts
@@ -25,7 +25,11 @@ export class Register<T> {
 				const fn = item;
 				result[key] = fn();
 			} else if (typeof item === 'object') {
-				result[key] = this.traverseObject(item, {});
+				if(item instanceof Register) {
+					result[key] = item.build();
+				} else {
+					result[key] = this.traverseObject(item, {});
+				}
 			} else {
 				result[key] = item;
 			}