+++ src/generators/geographic/index.ts
@@ -1,6 +1,11 @@
 import address from './address';
 import map from './map';
 
+export const Geo = {
+	address,
+	map
+};
+
 export default {
 	address,
 	map