+++ src/generators/internet/authentication.ts
@@ -1,4 +1,7 @@
-// export function JWT() {
+import { sign, SignOptions } from 'jsonwebtoken';
+import { fromFormat } from '../util/string';
 
-// }
-
+export function JWT(payload: string | object | Buffer = '', privateKey = '', options: SignOptions = {}): string {
+	if (!privateKey) privateKey = fromFormat('aAaAaAaAaAaAaAaA');
+	return sign(payload, privateKey, options);
+}