+++ src/generators/internet/http.ts
@@ -1,0 +1,19 @@
+interface HttpRequest {
+    headers: { [key: string]: string | number };
+    method: 'GET' | 'POST' | 'DELETE' | 'PUT';
+    params: string[];
+    query: string;
+}
+
+interface HttpResponse {
+    headers: { [key: string]: string | number };
+    status: number;
+}
+
+// export function request() {
+
+// }
+
+// export function response() {
+    
+// }