+++ test/core/contingency.spec.ts
@@ -1,6 +1,7 @@
 import { expect } from "chai";
 import { Contingency } from "../../src/core/contingency";
 import { Register } from "../../src/core/register";
+import { int } from './../../src/generators/math/number'
 
 describe('Contingency Funciton Test', () => {
     it('Should set a boolean contingent to a persons age', () => {
@@ -47,5 +48,23 @@ describe('Contingency Funciton Test', () => {
         expect(peopleResult).to.be.deep.equal({
             person: { age: 17 }, analytis: { track: false }
         });
+    });
+
+    it('Should be contingent to another generated value further up the chain', () => {
+        // create the analytics register
+        const analytics = {
+            id: int
+        }
+        const analyticsRegister = new Register(analytics);
+        expect(typeof analyticsRegister.build().id).to.be.equal('number');
+
+        // create the person register
+        const person = {
+            age: int,
+            analytics: analytics
+        };
+
+        const personRegister = new Register(person);
+        expect(typeof personRegister.build().age).to.be.equal('number');
     });
 });