+++ test/core/register.spec.ts
@@ -2,8 +2,8 @@ import { expect } from "chai";
 import { Register } from "../../src/core/register";
 
 const obj = {
-    'test': () => { return 1 },
-    'test-2': () => { return 'hello' },
+    'test': () => 1,
+    'test-2': () => 'hello',
 }
 
 describe('Register tests', () => {
@@ -12,12 +12,22 @@ describe('Register tests', () => {
         expect(register).to.be.instanceOf(Register);
     });
 
-    it('Should create a dictionary with multiple functions', () => {
+    it('Should create a register with multiple functions', () => {
         const register = new Register(obj);
         expect(typeof register.get('test')()).to.be.equal('number')
         expect(register.get('test')()).to.be.equal(1)
         expect(typeof register.get('test-2')()).to.be.equal('string')
         expect(register.get('test-2')()).to.be.equal('hello')
+    });
+
+    it('Should build a register which returns a fixed value instead of a returning function', () => {
+        const staticValues = {
+            name: 'hello-world'
+        };
+        const staticRegister = new Register(staticValues);
+        expect(staticRegister.build()).to.deep.equal({
+            name: 'hello-world'
+        });
     });
 
     it('Should create a dictionary and replace a function after a call', () => {
@@ -31,6 +41,29 @@ describe('Register tests', () => {
 
         register.add('password', () => '********');
         expect(register.get('password')()).to.be.equal('********');
-        
+    });
+
+    it('Should build an instance of the register', () => {
+        const register = new Register(obj);
+        const result = register.build();
+        expect(result.test).to.be.equal(1);
+        expect(result["test-2"]).to.be.equal('hello');
+    });
+
+    it('Should build a register with multiple objects', () => {
+        const obj2 = {
+            'hello': obj,
+            'world': () => 'world'
+        }
+
+        const register = new Register(obj2);
+        const result = register.build();
+        expect(result).to.deep.equal({
+            'hello': {
+                'test': 1,
+                'test-2': 'hello'
+            },
+            'world': 'world'
+        });
     });
 });